+++ includes/Loader.php
@@ -27,7 +27,7 @@ final class Loader
 			Base\Localization::class,
 			Core\Connection::class,
 			Core\MigrateActions::class,
-			Core\AdminBar::class,
+			Core\Admin::class,
 			Core\Settings::class,
 		];